start of collections refacotr
