+++ app/api/deck_lists.py
@@ -1,5 +1,5 @@
 from flask import Blueprint, request
-from app.models import DeckList, Deck, db
+from app.models import Collection, Deck, db
 from app.forms import NewDeckListForm, EditDeckListForm
 from .auth_routes import validation_errors_to_error_messages
 from app.awsupload import (
@@ -11,13 +11,13 @@ deck_list_routes = Blueprint('decklists', __name__)
 
 @deck_list_routes.route('/<int:id>/')
 def get_full_deck_list_detail(id):
-    decklist = DeckList.query.get(int(id))
+    decklist = Collection.query.get(int(id))
     return decklist.to_dict()
 
 
 @deck_list_routes.route('/cards/<int:id>/')
 def get_all_cards_from_list(id):
-    decklist = DeckList.query.get(int(id))
+    decklist = Collection.query.get(int(id))
     return {'cards': decklist.get_cards()}
 
 
@@ -44,7 +44,7 @@ def create_deck_list():
 
         url = upload["url"]
     if form.validate_on_submit():
-        newDeckList = DeckList(
+        newDeckList = Collection(
             title=form.data['title'], cover_photo_url=url, user_id=form.data['user_id'], has_image=form.data['has_image'])
         db.session.add(newDeckList)
         db.session.commit()
@@ -56,7 +56,7 @@ def create_deck_list():
 @deck_list_routes.route('/<int:id>/',  methods=["PUT"])
 def edit_deck_list(id):
     form = EditDeckListForm()
-    deckListToEdit = DeckList.query.get(int(id))
+    deckListToEdit = Collection.query.get(int(id))
     form['csrf_token'].data = request.cookies['csrf_token']
     url = 'No Photo'
 
@@ -91,7 +91,7 @@ def edit_deck_list(id):
 
 @deck_list_routes.route('/<int:id>/add/<int:deck_id>/', methods=["PATCH"])
 def add_deck_to_list(id, deck_id):
-    deckList = DeckList.query.get(int(id))
+    deckList = Collection.query.get(int(id))
     deck = Deck.query.get(int(deck_id))
     if deck:
         response = deckList.add_deck(deck)
@@ -105,7 +105,7 @@ def add_deck_to_list(id, deck_id):
 
 @deck_list_routes.route('/<int:id>/remove/<int:deck_id>/', methods=["PATCH"])
 def remove_deck_from_list(id, deck_id):
-    deckList = DeckList.query.get(int(id))
+    deckList = Collection.query.get(int(id))
     deck = Deck.query.get(int(deck_id))
     if deck:
         response = deckList.remove_deck(deck)
@@ -119,7 +119,7 @@ def remove_deck_from_list(id, deck_id):
 
 @deck_list_routes.route('/<int:id>/',  methods=["DELETE"])
 def delete_list(id):
-    listToDelte = DeckList.query.get(int(id))
+    listToDelte = Collection.query.get(int(id))
     if listToDelte:
         db.session.delete(listToDelte)
         db.session.commit()
