+++ app/models/__init__.py
@@ -1,7 +1,8 @@
 from .db import db
 from .user import User
 from .categorites import Category
-from .decks import Deck, DeckList
+from .decks import Deck
 from .cards import Card
+from .collections import Collection
 from .joined_decks import AddedDeck, CompletedDeck
 from .badges import Badge, UserBadge
