+++ app/models/joined_decks.py
@@ -17,7 +17,7 @@ class AddedDeck(db.Model):
     __tablename__ = 'added_decks'
     deck_id = db.Column(db.Integer, db.ForeignKey(
         'decks.id'), primary_key=True)
-    deck_list_id = db.Column(db.Integer, db.ForeignKey(
-        'deck_lists.id'), primary_key=True)
+    collection_id = db.Column(db.Integer, db.ForeignKey(
+        'collections.id'), primary_key=True)
     created_on = db.Column(db.DateTime, nullable=False,
                            default=datetime.utcnow)
