+++ app/seeds/__init__.py
@@ -3,7 +3,7 @@ from .users import seed_users, undo_users
 from .categories import seed_categories, undo_categories
 from .decks import seed_decks, undo_decks
 from .cards import seed_cards, undo_cards
-from .deck_lists import seed_deck_lists, undo_deck_lists
+from .collections import seed_collections, undo_collections
 from .badges import seed_badges, undo_badges
 from .user_badges import seed_user_badges, undo_user_badges
 
@@ -22,7 +22,7 @@ def seed():
     seed_categories()
     seed_decks()
     seed_cards()
-    seed_deck_lists()
+    seed_collections()
 
     # Add other seed functions here
 
@@ -36,5 +36,5 @@ def undo():
     undo_categories()
     undo_decks()
     undo_cards()
-    undo_deck_lists()
+    undo_collections()
     # Add other undo functions here
